Move CompressionCodec to validation package
