+++ java/configuration/src/main/java/sleeper/configuration/Utils.java
@@ -19,7 +19,6 @@ import com.google.common.collect.Sets;
 import org.apache.commons.lang3.EnumUtils;
 
 import sleeper.configuration.properties.SleeperPropertyValues;
-import sleeper.configuration.properties.table.CompressionCodec;
 import sleeper.configuration.properties.validation.EmrInstanceArchitecture;
 
 import java.util.List;
@@ -93,10 +92,6 @@ public class Utils {
 
     public static boolean isValidFadvise(String fadvise) {
         return VALID_FADVISE_VALUES.contains(fadvise);
-    }
-
-    public static boolean isValidCompressionCodec(String codec) {
-        return EnumUtils.isValidEnumIgnoreCase(CompressionCodec.class, codec);
     }
 
     public static boolean isValidLogRetention(String logRetention) {
